Refactor movement logic for chess pieces, introduce direction enums for Bishop, King, Knight, Rook, and Queen. Create FEN class but need to repair the tests. Line Coverage figures: 99%
