+++ src/test/kotlin/hwr/oop/MovementTests.kt
@@ -7,8 +7,6 @@ import hwr.oop.figures.Queen
 import hwr.oop.figures.Rook
 import io.kotest.core.spec.style.AnnotationSpec
 import org.assertj.core.api.Assertions.assertThat
-import java.io.ByteArrayOutputStream
-import java.io.PrintStream
 
 class MovementTests : AnnotationSpec() {
